chemstreamlit_js/requirements.txt
streamlit==0.72.0
numpy==1.19.2
Pillow==8.0.1
rdkit
scikit_learn==0.23.2
